backend/requirements.txt
fastapi>=0.115,<1.0
uvicorn[standard]>=0.30
sqlalchemy>=2.0,<3.0
aiosqlite>=0.20
alembic>=1.14
pydantic>=2.10
pydantic-settings>=2.6
httpx>=0.28
python-multipart>=0.0.18
aiofiles>=24.1
playwright>=1.49
faster-whisper>=1.1
apscheduler>=4.0.0a5
websockets>=14.0
# Upstream pipeline dependencies
anthropic>=0.40
moviepy>=2.0
Pillow>=10.0
edge-tts>=6.1
python-dotenv>=1.0
